+++ src/main/java/uk/gov/hmcts/reform/pcs/notify/repository/NotificationRepository.java
@@ -6,7 +6,7 @@ import uk.gov.hmcts.reform.pcs.notify.entities.CaseNotification;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface NotificationRepository extends JpaRepository<CaseNotification, Long> {
+public interface NotificationRepository extends JpaRepository<CaseNotification, Integer> {
     /**
      * Find a case notification by its provider notification ID.
      *
