+++ src/main/java/uk/gov/hmcts/reform/pcs/testingsupport/endpoint/NotifyController.java
@@ -41,7 +41,7 @@ public class NotifyController {
     public ResponseEntity<List<EmailNotificationResponse>> sendDefendantResponseEmails(
         @RequestHeader(value = AUTHORIZATION, defaultValue = "DummyId") String authorisation,
         @RequestHeader(value = "ServiceAuthorization") String serviceAuthorization,
-        @RequestParam Long defendantResponseId) {
+        @RequestParam Integer defendantResponseId) {
 
         log.info("Received request to send all defendant response emails for {}", defendantResponseId);
         // temporary endpoint to test sending emails
